+++ CHANGELOG.rst
@@ -2,6 +2,19 @@ This document contains the ``fslpy`` release history in reverse chronological
 order.
 
 
+3.15.4 (Monday 27th November 2023)
+----------------------------------
+
+
+Added
+^^^^^
+
+* New `silent` option to the :func:`.run` function = passing ``silent=True`` is
+  equivalent to passing ``log={'tee':False}`` (!428).
+* New `prefix` option to the :func:`.tempdir` function, which is passed through
+  to ``tempfile.mkdtemp`` (!429).
+
+
 3.15.3 (Thursday 16th November 2023)
 ------------------------------------